pyproject.toml
[tool.poetry]
name = "chatgpt-cbs-opendata-plugin"
version = "0.1.0"
description = ""
authors = ["Yeb Havinga <yhavinga@gmail.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.0.0"
cbsodata = "^1.3.4"
fastapi = "^0.95.0"
pydantic = "^1.10.7"
openai = "^0.27.4"
tiktoken = "^0.3.3"
pytest = "^7.3.0"
asyncpg = "^0.27.0"
tenacity = "^8.2.2"
backoff = "^2.2.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.23.3"
uvicorn = "^0.21.1"
python-dotenv = "^1.0.0"
safety = "^2.3.5"
faiss-cpu = "^1.7.3"
duckdb = "^0.7.1"
restrictedpython = "^6.0"
diskcache = "^5.4.0"

[tool.poetry.scripts]
start = "server.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
